Restructured project. * file structure is updated inside users-api/readme.md * and, whole ecommerce readme in the project root. * removed google auth as it's now intended to go inside seperate auth-server-api.
