+++ users-api/cmd/app/handlers.go
@@ -1,8 +1,8 @@
 package app
 
 import (
-	"github.com/ashtishad/ecommerce/domain"
-	"github.com/ashtishad/ecommerce/service"
+	"github.com/ashtishad/ecommerce/users-api/internal/domain"
+	"github.com/ashtishad/ecommerce/users-api/internal/service"
 	"github.com/gin-gonic/gin"
 	"log"
 	"net/http"
